requirements.txt
python-dateutil
PyYAML
requests
hjson
PyJWT
mock
pylint
responses
pytest
pytest-cover
pylint2junit
twine
black
tox
sphinx
sphinx-autodoc-typehints
sphinx-rtd-theme
pandas
jupyter-sphinx-theme